Space Knowledge Base has inline Callouts list
